pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]

[project]
name = "ai-venture-studio"
version = "0.1.0"
description = "Autonomous AI Venture Studio — LangGraph 10-phase pipeline with Postgres and Aiogram approval"
readme = "docs/PHASES_INDEX.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.13.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=3.0.0",
    "langchain-core>=0.3.0",
    "psycopg[binary,pool]>=3.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
